cli: make WASI an optional build tag (wago_wasi), out of default build
